mtr: math fixes, remove geometric mean
First, geometric mean is calculated wrong in the original source code.
There is overflow when calculated it the correct way and math/big doesn't
have power functions to make it accurate and I don't want to implement newton
estimation.
Second, geometric mean is meant to enable comparison between a collection of disparate
things. It shows central tendency of multiple variables. A geometric mean on one variable (packet timings) is essentially the same
calculation as a normal average of all those values.

Let's do the math:
     The 6th root of 3*4*3*2*4*3   ~= 3.086
                     3+4+3+2+4+3/6 ~= 3.167

The useless-ness of geometric mean on one variable can be seen from calculating it on
other numbers as well.